Move some input related code out of application files
- Move some enums and a function to input.{h, cpp}
- Also remove sCharMap and CursorMode since they don't
appear to be used
